Separate incremental merge behavior from Foundry orchestration
The Foundry runner no longer owns the incremental planning helpers directly. Moving that behavior into its own app boundary keeps RunFoundry focused on orchestration while preserving the current cache-selection semantics.

Constraint: Keep the first refactor behavior-preserving and package-local before promoting reusable abstractions

Confidence: high

Scope-risk: narrow

Directive: Preserve the any-ok-wins cache heuristic unless Foundry ordering semantics are explicitly revisited

Tested: go test ./...

Not-tested: ./dev verify lint remains blocked by existing Go 1.25 linter vs Go 1.26 source-processing mismatch
